requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
Authlib==1.3.2
backoff==2.2.1
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cryptography==39.0.2
cymem==2.0.8
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.9.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
greenlet==3.1.1
groq==0.12.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.3.3
keras-models==0.0.7
Keras-Preprocessing==1.1.2
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.19
langchain-experimental==0.3.3
langchain-groq==0.2.1
langchain-text-splitters==0.3.0
langcodes==3.4.0
langsmith==0.1.137
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql==0.0.3
mysqlclient==2.2.6
namex==0.0.8
narwhals==1.10.0
networkx==3.4.2
numpy==1.26.0
oauthlib==3.2.2
opencage==2.1.0
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.10
packaging==24.0
pandas==2.1.1
pathlib==1.0.1
phonenumbers==8.13.7
pillow==10.3.0
preshed==3.0.9
propcache==0.2.0
protobuf==4.25.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.0
pyOpenSSL==23.0.0
pyparsing==3.2.0
pypdf==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.28.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.16.0
thinc==8.2.4
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.46.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.3
watchdog==5.0.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.16.0
